implemented config file loading

+++ src/faketorio-clean.lua
@@ -1,6 +1,11 @@
 if not faketorio then faketorio = {} end
 
 function faketorio.clean()
+    local attr = faketorio.lfs.attributes("target")
+    if (attr == nil) then
+        return
+    end
+
     faketorio.delete_dir("target")
     faketorio.lfs.rmdir("target")
 end
